feat(director): readable full-month dates across all surfaces
Standardize every date display on the director side to a full month
name format — "August 7, 2026" (and "August 7, 2026, 3:45 PM" for
timestamps) — replacing raw ISO strings (2026-08-07) and numeric
locale dates (8/7/2026).

- Add canonical fmtDate/fmtDateTime helpers in Charts.tsx (accept
  ms | ISO string | Date | null/"—"; ISO date-only parsed at local
  midnight to avoid TZ day-shift; sentinels pass through).
- Add fmtWhenExact in enrichment/helpers.tsx for hover-title tooltips.
- Route all local date helpers and inline toLocale* calls through the
  full-month format; wrap raw ISO field renders (scheduled, lastBooking,
  etc.) with fmtDate.
- Leave compact-by-design displays untouched: chart-axis M/D labels,
  the week strip, the weekday calendar cell, short-month event logs,
  and 24h scheduled_time strings.

Presentation-layer only; Convex still returns raw ISO.

+++ app/(director-panel)/director/components/tabs/TabVehicleConfigs.tsx
@@ -21,7 +21,7 @@ import { DirectorSessionCtx } from '../DirectorSessionCtx'
 
 function fmtDate(ts?: number): string {
   if (!ts) return '—'
-  return new Date(ts).toLocaleDateString('en-US', { month:'short', day:'numeric', year:'numeric' })
+  return new Date(ts).toLocaleDateString('en-US', { month:'long', day:'numeric', year:'numeric' })
 }
 
 function ageLabel(ts?: number): string {
